+++ react/RecommendationShelf.tsx
@@ -26,6 +26,23 @@ defineMessages({
     id: 'admin/editor.recommendation-shelf.display-title',
     defaultMessage: 'Display Title',
   },
+  itemsContext: {
+    id: 'admin/editor.recommendation-shelf.items-context',
+    defaultMessage: 'Items Context',
+  },
+  itemsContextDescription: {
+    id: 'admin/editor.recommendation-shelf.items-context.description',
+    defaultMessage:
+      'From where to retrieve the items that will be used as context in the Recommendation request',
+  },
+  itemsContextCart: {
+    id: 'admin/editor.recommendation-shelf.items-context.cart',
+    defaultMessage: 'Cart',
+  },
+  itemsContextPdp: {
+    id: 'admin/editor.recommendation-shelf.items-context.pdp',
+    defaultMessage: 'Product Page',
+  },
 })
 
 type Props = {
@@ -33,6 +50,7 @@ type Props = {
   title?: string
   recommendationType?: RecommendationType // Deprecated, use campaignVrn instead
   displayTitle: boolean
+  itemsContext: ItemContextType[]
 }
 
 const RecommendationShelf: StorefrontFunctionComponent<Props> = ({
@@ -40,6 +58,7 @@ const RecommendationShelf: StorefrontFunctionComponent<Props> = ({
   title,
   recommendationType,
   displayTitle,
+  itemsContext,
 }) => {
   return (
     <RecommendationShelfErrorBoundary>
@@ -48,6 +67,7 @@ const RecommendationShelf: StorefrontFunctionComponent<Props> = ({
         title={title}
         recommendationType={recommendationType}
         displayTitle={displayTitle}
+        itemsContext={itemsContext}
       />
     </RecommendationShelfErrorBoundary>
   )
@@ -70,6 +90,21 @@ RecommendationShelf.schema = {
       title: 'admin/editor.recommendation-shelf.display-title',
       type: 'boolean',
       default: true,
+    },
+    itemsContext: {
+      title: 'admin/editor.recommendation-shelf.items-context',
+      description:
+        'admin/editor.recommendation-shelf.items-context.description',
+      type: 'array',
+      items: {
+        type: 'string',
+        enum: ['PDP', 'CART'],
+        enumNames: [
+          'admin/editor.recommendation-shelf.items-context.pdp',
+          'admin/editor.recommendation-shelf.items-context.cart',
+        ],
+      },
+      default: ['PDP'],
     },
   },
 }